refactor(io): CodeRabbit items — isEnoent layering + moveToArchive inline + buildArtifactPath + path.posix.join (#384)
1. isEnoent moved from workspace-io.ts → safe.ts (lower layer).
   workspace-io + json.ts + journal-io now import from safe.ts.
   No re-export — CLAUDE.md prohibits.

2. moveToArchive wrapper deleted from optimizationPass.ts.
   archiveTopic(slug, workspaceRoot) called directly in both
   executeMergePlans and executeArchives.

3. buildArtifactPath (naming.ts) now used by:
   - presentHtml.ts: dropped slugify + ensureWorkspaceDir imports
   - chart.ts: dropped slugify + ensureWorkspaceDir imports
   Both files no longer construct filenames manually.

4. dailyPass.ts:550 last template literal slash →
   path.posix.join(WORKSPACE_DIRS.chat, ...).
